+++ dust-lang/src/instruction/less.rs
@@ -28,7 +28,7 @@ impl From<&Instruction> for Less {
 impl From<Less> for Instruction {
     fn from(less: Less) -> Self {
         let operation = Operation::LESS;
-        let a_field = less.comparator as u16;
+        let a_field = less.comparator as usize;
         let Address {
             index: b_field,
             memory: b_memory_kind,
